revert: drop DPNS test relocation (split to separate PR)
Reverts the following commits, moved to a dedicated PR off v3.1-dev so #3711
can stay focused on the V0/V1 wire compatibility fix:

  09df598f38  test(rs-sdk): make DPNS network tests honor tests/.env DASH_SDK_PLATFORM_* vars
  9745211854  refactor(rs-sdk): move DPNS network tests from src/ to tests/, drop band-aid endpoint helper
  23ee2ae13a  test(rs-sdk): wire relocated DPNS tests through Config::setup_api for local Core RPC quorum lookup
  23b3a4aaad  test(rs-sdk): loosen test_dpns_queries_from_docs to smoke-test shape (chain-agnostic)

The contested_resource vector regeneration (f384ca7c3d) is independent and
stays on this PR.
